requirements.txt
jinja2
pyyaml
argparse
